+++ starsky/starskysynccli/Program.cs
@@ -56,6 +56,7 @@ namespace starskysynccli
             if (new ArgsHelper(appSettings).GetThumbnail(args))
             {
 	            var storage = startupHelper.SubPathStorage();
+	            var thumbnailStorage = startupHelper.ThumbnailStorage();
 
 				var isFolderOrFile = storage.IsFolderOrFile(subpath);
 
@@ -65,11 +66,11 @@ namespace starskysynccli
                 {
                     // If single file => create thumbnail
 	                var fileHash = new FileHash(storage).GetHashCode(subpath);
-                    new Thumbnail(storage).CreateThumb(subpath,fileHash); // <= this uses subpath
+                    new Thumbnail(storage,thumbnailStorage).CreateThumb(subpath,fileHash); // <= this uses subPath
                 }
                 else
                 {
-	                new Thumbnail(storage).CreateThumb(subpath);
+	                new Thumbnail(storage, thumbnailStorage).CreateThumb(subpath);
                 }
                 
                 Console.WriteLine("Thumbnail Done!");