+++ history.md
@@ -23,7 +23,9 @@
 - [ ]	(feature) _Back-end_ Injection framework implemented
 - [ ] 	(rename) _Back-end_ Feature renaming and docs updates
 - [ ] 	(feature) _Back-end support for RAW that is not Sony for example Nikon .NEF
+- [ ]   (feature) tiff, arw:sony, dng:adobe, nef:nikon, raf:fuji, cr2:canon, crw:canon, orf:olympus, rw2:panasonic, pef:pentax, x3f:sigma
 - [ ]   (bugfix) _Back-end_ allow underscore import/upload
+- [ ]   (bugfix) _Front-end_ Download selection thumbnail right extension suggestion
 
 # version 0.1.17 - 2020-03-07
 - [x]  (feature) _Front-end_ DateTime editing in detailView